STYLE-NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE fix(#191)

+++ core/src/com/pspace/ifs/ksan/osd/EventObject.java
@@ -182,10 +182,12 @@ class MoveObjectCallback implements MQCallback {
 				String ecDir = KsanUtils.makeECDirectoryPath(sourceDiskPath, objId);
 				File dir = new File(ecDir);
 				File[] ecFiles = dir.listFiles();
-				for (int i = 0; i < ecFiles.length; i++) {
-					if (ecFiles[i].getName().startsWith(Constants.POINT)) {
-						if (ecFiles[i].getName().charAt(ecFiles[i].getName().length() - 2) == Constants.CHAR_POINT) {
-							ecFiles[i].delete();
+				if (ecFiles != null) {
+					for (int i = 0; i < ecFiles.length; i++) {
+						if (ecFiles[i].getName().startsWith(Constants.POINT)) {
+							if (ecFiles[i].getName().charAt(ecFiles[i].getName().length() - 2) == Constants.CHAR_POINT) {
+								ecFiles[i].delete();
+							}
 						}
 					}
 				}
@@ -482,10 +484,12 @@ class GetAttrObjectCallBack implements MQCallback {
 				String ecDir = KsanUtils.makeECDirectoryPath(diskPath, objId);
 				File dir = new File(ecDir);
 				File[] ecFiles = dir.listFiles();
-				for (int i = 0; i < ecFiles.length; i++) {
-					if (ecFiles[i].getName().startsWith(Constants.POINT)) {
-						if (ecFiles[i].getName().charAt(ecFiles[i].getName().length() - 2) == Constants.CHAR_POINT) {
-							ecFiles[i].delete();
+				if (ecFiles != null) {
+					for (int i = 0; i < ecFiles.length; i++) {
+						if (ecFiles[i].getName().startsWith(Constants.POINT)) {
+							if (ecFiles[i].getName().charAt(ecFiles[i].getName().length() - 2) == Constants.CHAR_POINT) {
+								ecFiles[i].delete();
+							}
 						}
 					}
 				}
@@ -658,10 +662,12 @@ class CopyObjectCallback implements MQCallback {
 				String ecDir = KsanUtils.makeECDirectoryPath(sourceDiskPath, objId);
 				File dir = new File(ecDir);
 				File[] ecFiles = dir.listFiles();
-				for (int i = 0; i < ecFiles.length; i++) {
-					if (ecFiles[i].getName().startsWith(Constants.POINT)) {
-						if (ecFiles[i].getName().charAt(ecFiles[i].getName().length() - 2) == Constants.CHAR_POINT) {
-							ecFiles[i].delete();
+				if (ecFile != null) {
+					for (int i = 0; i < ecFiles.length; i++) {
+						if (ecFiles[i].getName().startsWith(Constants.POINT)) {
+							if (ecFiles[i].getName().charAt(ecFiles[i].getName().length() - 2) == Constants.CHAR_POINT) {
+								ecFiles[i].delete();
+							}
 						}
 					}
 				}
